requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
